fix: move error parsing methods
