+++ src/common/hooks/useEndorsementChain/useEndorsementChain.ts
@@ -7,6 +7,7 @@ import { mergeTransfers } from "./helpers";
 import { fetchTokenTransfers } from "./fetchTokenTransfer";
 import { getEndorsementChain } from "./retrieveEndorsementChain";
 import { retrieveTitleEscrowAddressOnFactory } from "../useTitleEscrowContract";
+import { getErrorMessage } from "../../utils/errorParsing";
 
 export const useEndorsementChain = (
   tokenRegistryAddress: string,
@@ -47,31 +48,4 @@ export const useEndorsementChain = (
   }, [fetchEndorsementChain]);
 
   return { endorsementChain, pending, error };
-};
-
-type ErrorWithMessage = {
-  message: string;
-};
-
-export const isErrorWithMessage = (error: unknown): error is ErrorWithMessage => {
-  return (
-    typeof error === "object" &&
-    error !== null &&
-    "message" in error &&
-    typeof (error as Record<string, unknown>).message === "string"
-  );
-};
-
-export const toErrorWithMessage = (maybeError: unknown): ErrorWithMessage => {
-  if (isErrorWithMessage(maybeError)) return maybeError;
-
-  try {
-    return new Error(JSON.stringify(maybeError));
-  } catch {
-    return new Error(String(maybeError));
-  }
-};
-
-export const getErrorMessage = (error: unknown): string => {
-  return toErrorWithMessage(error).message;
 };
