+++ src/common/utils/errorParsing.ts
@@ -1,0 +1,26 @@
+type ErrorWithMessage = {
+  message: string;
+};
+
+export const isErrorWithMessage = (error: unknown): error is ErrorWithMessage => {
+  return (
+    typeof error === "object" &&
+    error !== null &&
+    "message" in error &&
+    typeof (error as Record<string, unknown>).message === "string"
+  );
+};
+
+export const toErrorWithMessage = (maybeError: unknown): ErrorWithMessage => {
+  if (isErrorWithMessage(maybeError)) return maybeError;
+
+  try {
+    return new Error(JSON.stringify(maybeError));
+  } catch {
+    return new Error(String(maybeError));
+  }
+};
+
+export const getErrorMessage = (error: unknown): string => {
+  return toErrorWithMessage(error).message;
+};
